pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zero-mind"
version = "0.1.0"
description = "Unified ARC reasoning research codebase"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.3.0",
  "pyyaml>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
